arp/rep_failure_notification_email
