+++ modules/accredited_representative_portal/lib/accredited_representative_portal/engine.rb
@@ -15,6 +15,11 @@ module AccreditedRepresentativePortal
 
     config.generators.api_only = true
 
+    # Make sure Rails autoloads lib/ properly
+    initializer :append_lib_to_autoload_paths do |_app|
+      ActiveSupport::Dependencies.autoload_paths << root.join('lib')
+    end
+
     # So that the app-wide migration command notices our engine's migrations.
     initializer :append_migrations do |app|
       unless app.root.to_s.match? root.to_s
@@ -29,14 +34,17 @@ module AccreditedRepresentativePortal
       FactoryBot.definition_file_paths << File.expand_path('../../spec/factories', __dir__) if defined?(FactoryBot)
     end
 
-    initializer 'accredited_representative_portal.benefits_intake.register_handler' do |app|
+    initializer 'accredited_representative_portal.benefits_intake.register_handler',
+                after: :append_lib_to_autoload_paths do |app|
       app.config.to_prepare do
         require 'lighthouse/benefits_intake/sidekiq/submission_status_job'
         require 'burials/benefits_intake/submission_handler'
 
         # Register our Claims Benefits Intake Submission Handler
-        ::BenefitsIntake::SubmissionStatusJob.register_handler(SavedClaim::BenefitsIntake::DependencyClaim::PROPER_FORM_ID,
-                                                               AccreditedRepresentativePortal::BenefitsIntake::SubmissionHandler)
+        ::BenefitsIntake::SubmissionStatusJob.register_handler(
+          AccreditedRepresentativePortal::SavedClaim::BenefitsIntake::DependencyClaim::PROPER_FORM_ID,
+          AccreditedRepresentativePortal::SubmissionHandler
+        )
       end
     end
   end
