+++ modules/accredited_representative_portal/spec/lib/notification_email_spec.rb
@@ -25,3 +25,5 @@ RSpec.describe AccreditedRepresentativePortal::NotificationEmail do
     end
   end
 end
+
+# no factory class
